chore: refactor DataToPoint
